[REFACTOR] 파일 네이밍 수정 및 네이밍 확인 패키지 추가
