+++ package.json
@@ -6,7 +6,7 @@
     "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
+    "lint:fix": "next lint --fix",
     "prisma:generate": "dotenv -e .env.local -- npx prisma generate",
     "prisma:migrate:dev": "dotenv -e .env.local -- prisma migrate dev --name"
   },
@@ -38,9 +38,11 @@
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "@typescript-eslint/eslint-plugin": "^8.25.0",
     "dotenv-cli": "^8.0.0",
     "eslint": "^9",
     "eslint-config-next": "15.1.6",
+    "eslint-plugin-check-file": "^3.1.0",
     "postcss": "^8",
     "prisma": "^6.3.1",
     "tailwindcss": "^3.4.1",
