+++ src/app/register/page.tsx
@@ -1,4 +1,4 @@
-import RegisterForm from '@/components/auth/register-form'
+import RegisterForm from '@/components/auth/RegisterForm'
 
 export const metadata = {
     title: '회원가입',
